Castings to support building on windows + windows CI (#20)

+++ src/types/direction.rs
@@ -1,3 +1,5 @@
+// Allow clippy::unnecessary_cast since the cast is needed for when bindgen runs on windows. The enum variants get cast to i32 on windows.
+#![allow(clippy::unnecessary_cast)]
 use strum::FromRepr;
 
 use crate::{sys::*, Error, Result};
@@ -6,8 +8,8 @@ use crate::{sys::*, Error, Result};
 #[derive(Copy, Clone, Debug, FromRepr, PartialEq, Eq)]
 #[repr(u32)]
 pub enum Direction {
-    RX = bladerf_direction_BLADERF_RX,
-    TX = bladerf_direction_BLADERF_TX,
+    RX = bladerf_direction_BLADERF_RX as u32,
+    TX = bladerf_direction_BLADERF_TX as u32,
 }
 
 impl From<Direction> for bladerf_direction {
@@ -20,7 +22,7 @@ impl TryFrom<bladerf_direction> for Direction {
     type Error = Error;
 
     fn try_from(value: bladerf_direction) -> Result<Self> {
-        Self::from_repr(value)
+        Self::from_repr(value as u32)
             .ok_or_else(|| Error::msg(format!("Invalid Direction value: {value}")))
     }
 }
